fix: clear error description and simplify message
No longer appends error description to the tmp_facilities table, clears - sets to null, all error descriptions for attempting to update the facilities table. Uses the simplified message used in the message box. No need to include sql as this found in table if tis modifying, or it is adding or deleting.